+++ go/entity/userMood.go
@@ -8,14 +8,14 @@ import (
 type UserMood struct {
 	ID           string
 	UserID       string
-	Mood         int
+	Mood         float64
 	CreationDate time.Time
 }
 
 type UserMoodPayload struct {
 	ID           string
 	UserID       string
-	Mood         int
+	Mood         float64
 	CreationDate time.Time
 }
 
@@ -26,8 +26,8 @@ type UserMoodFilters struct {
 // ValidateMood checks if given mood is a valid number in the specified range.
 func (ump *UserMoodPayload) ValidateMood() error {
 	validMood := false
-	minValue := 0
-	maxValue := 10
+	minValue := 0.0
+	maxValue := 10.0
 	if ump.Mood >= minValue && ump.Mood <= maxValue {
 		validMood = true
 	}
